# Sentmap refactorisation