… lib/sentmap.c