@kazuho 2020-04-04
Thank you for the PR.

Am I correct to assume that the benchmark was conducted on a low-latency, high-bandwidth network, but still sees some congestions? I would like to reproduce the numbers you saw.