@AloysAugustin 2020-04-04
Yes, the tests were performed on a very low latency, high throughput network.
We tested with our VPP integration where this change allowed us to go from 5.9 to 6.3Gbit/s with 2000 bytes MTU.
The improvement can also be seen with qperf (https://github.com/rbruenig/qperf), which is probably simpler to use. The upstream version is not currently compatible with quicly master, so for now you should use my fork if you would like to test it: https://github.com/AloysAugustin/qperf.

Here are the results I obtained with qperf on a skylake machine running Ubuntu 18.04, with all default options:
Quicly master:
```
~/qperf/build$ ./qperf -c localhost
running client with host=localhost and runtime=10s
connection establishment time: 5ms
time to first byte: 5ms
second 0: 2.02 gbit/s (271150745 bytes received)
second 1: 2.083 gbit/s (279641717 bytes received)
second 2: 2.141 gbit/s (287408614 bytes received)
second 3: 2.156 gbit/s (289314994 bytes received)
second 4: 2.145 gbit/s (287874893 bytes received)
second 5: 2.156 gbit/s (289403653 bytes received)
second 6: 2.181 gbit/s (292782550 bytes received)
second 7: 2.181 gbit/s (292740093 bytes received)
second 8: 2.165 gbit/s (290612150 bytes received)
second 9: 2.105 gbit/s (282572043 bytes received)
connection closed
```
With this change:
```
~/qperf/build$ ./qperf -c localhost
running client with host=localhost and runtime=10s
connection establishment time: 5ms
time to first byte: 5ms
second 0: 1021 mbit/s (133791763 bytes received)
second 1: 2.271 gbit/s (304856781 bytes received)
second 2: 2.312 gbit/s (310342380 bytes received)
second 3: 2.297 gbit/s (308330612 bytes received)
second 4: 2.308 gbit/s (309803357 bytes received)
second 5: 2.237 gbit/s (300284822 bytes received)
second 6: 2.301 gbit/s (308882886 bytes received)
second 7: 2.349 gbit/s (315291804 bytes received)
second 8: 2.393 gbit/s (321175290 bytes received)
second 9: 2.398 gbit/s (321790661 bytes received)
connection closed
```