@AloysAugustin 2020-04-13
Thank you for taking the time to try this PR.

I ran more tests with and without #319, both with qperf and VPP.
What I noticed is that this PR makes little difference unless there are dropped packets in the test.
For instance with VPP, we reach 6.1 Gbps with the current master (which includes #319), and 6.3 with this PR on top of master, with ~100 packets dropped per second.
I think this is due to the fact that drops cause out-of-order deletes in the sentmap, which requires walking through the previous entries with the current implementation.

Overall I think this deserves more testing at this point, to run tests while controlling more accurately the drop rate and maybe switch the congestion control algorithm to something that can sustain high throughputs even in the presence of drops. I will let you know if I get to spend some time on it.

As a side note, I also tried to keep the allocated packets in a freelist in order to reduce the amount of calls to malloc / free, but I did not observe a significant difference.